requirements.txt
mysql-connector-python==9.2.0
sodapy==2.2.0